refactor: move `config` folder into `src/common`
