+++ prisma/generate.ts
@@ -1,4 +1,4 @@
-import { dbConfig } from '~/utils/config/db';
+import { dbConfig } from '@/common/config/db';
 import { validateDBConfig } from '~/utils/db';
 import { runOrFail } from '../utils/helpers';
